+++ R/use_crs.R
@@ -2,8 +2,10 @@
 #'
 #' Set or return the coordinate system currently in use.
 #'
-#' If argument `crs` is NULL, the function returns the current value (which may be `NULL``).
+#' If argument `crs` is NULL, the function returns the current value (which may be `NULL`).
 #' @param crs provide PROJ string to set the value
+#' @return the current value of the in-use coordinate system (may be `NULL`), if
+#' `crs` is provided it is returned after being set as the in-use value
 #' @export
 #' @examples
 #' \dontrun{