+++ tests/testthat/test-qsc.R
@@ -1,4 +1,3 @@
-context("test-qsc")
 
 test_that("qsc projection works", {
   expect_silent(qsc())