+++ tests/testthat/test-texture.R
@@ -1,4 +1,3 @@
-context("test-texture")
 
 g <- raster::setExtent(raster::raster(matrix(1:12, 3)), raster::extent(0, 4, 0, 3))
 test_that("texturing works", {